autogen_trigger/requirements.txt
fastapi
uvicorn
python-dotenv
pyautogen